Move CacheManager to infrastructure package
