+++ webapi/src/main/scala/org/knora/webapi/slice/admin/api/service/StoreRestService.scala
@@ -12,7 +12,7 @@ import org.knora.webapi.config.AppConfig
 import org.knora.webapi.messages.store.triplestoremessages.RdfDataObject
 import org.knora.webapi.messages.util.KnoraSystemInstances.Users.SystemUser
 import org.knora.webapi.slice.admin.api.MessageResponse
-import org.knora.webapi.slice.admin.repo.service.CacheManager
+import org.knora.webapi.slice.infrastructure.CacheManager
 import org.knora.webapi.slice.ontology.repo.service.OntologyCache
 import org.knora.webapi.store.triplestore.api.TriplestoreService
