+++ webapi/src/test/scala/org/knora/webapi/slice/admin/repo/service/KnoraUserRepoLiveSpec.scala
@@ -38,6 +38,7 @@ import org.knora.webapi.slice.admin.domain.model.UserStatus
 import org.knora.webapi.slice.admin.domain.model.Username
 import org.knora.webapi.slice.admin.domain.service._
 import org.knora.webapi.slice.admin.repo.rdf.Vocabulary
+import org.knora.webapi.slice.infrastructure.CacheManager
 import org.knora.webapi.store.triplestore.api.TestTripleStore
 import org.knora.webapi.store.triplestore.api.TriplestoreService
 import org.knora.webapi.store.triplestore.api.TriplestoreService.Queries.Update
